[Monorail] Double the default API QPM.
The purpose of rate limits is to prevent one user from putting too much
load on the server such that other users would be impacted.  We don't
see that happening at 100 QPM, even though we have a lot of clients.
And, the overhead of managing requests for custom QPMs is a distraction.
So, let's try a higher default limit.

R=jeffcarp@chromium.org

Bug: monorail:4381
Change-Id: If04f6e18cc962083bc01c2258ce80d84b26903ff