crosskylabadmin: Add a testFixture
The new testFixture holds the top level API server and fake/mock objects
needed for all frontend tests. Collecting these in a struct will allow
us to flatten the existing unittests (which entails duplication of the
test setup boilerplate) and saner migration from a fake SwarmingClient
to a mock SwarmingClient.

BUG=None

Change-Id: Idc769c67e4906d721f9c7bfa993b00ba60578ceb