crosskylabadmin: Flatten tracker tests
Tracker unittests were getting too long to be readable and the multiple
levels of Convey() blocks required the reader to track local variable
names across irrelevant sibling Convey() blocks when reading a test.

Flatten the tests by reducing the depth of Convey() trees. Split up some
tests in the process. This also makes it easier to replace the use of
fake SwarmingClient with a mock.

BUG=None

Change-Id: Ifc27251a06ff87e6ad4a246aff6e27e055e6b111