crosskylabadmin: Add note about getIdleDuration
This got in the way of my refactoring, to reuse the Swarming query for
bot tasks for both the idle duration and DUT status diagnosis, since
getIdleDuration calls Swarming directly.

Not sure how I want to fix this right now or if I want to fix it right
now, so leaving a note.

R=pprabhu@chromium.org, tandrii@chromium.org

Change-Id: I09f2f1900bec2c4e4e72afba78d6f5f5ced26a41