[Findit] Create CQHiddenFlakeOccurrence to store occurrence groups of cq hidden flakes.
Because we expecting a large amount of hidden flake occurrences, so we will not store each and every such occurrences to data store. Instead, we will use one entity to store a group of occurrences if they:
  1. are detected within the same cron job,
  2. are for the same test, meaning they have the same step_ui_name and
    test_name,
  3. are on the same builder.

Full information of one occurrence in the group will be stored as a sample, while others will be consolidated and only save their distinctive information.

Bug: 896004
Change-Id: I95755a2ce9bd675c8a6b1fb3eed02d0a077bb823