Update patch set 7
Patch Set 7:

(5 comments)

Higher level question. Can the MetricsSink be attached to the state type (and other types) so that all the methods don't have to have it in the argument list.

MetricsSink is a fixture that form the environment for the state and other types. Usually, such fixtures are provided in the constructor and used from all the methods (dependency injection).

Patch-set: 7