Emit PRESUBMIT_FAILURE as fail_type for processed CQ events.
The False Rejects SLO is supposed to ignore CQ attempts that failed due to a
missing LGTM. Unfortunately, that information is not available in the CQ events
table. As a proxy for this information, this CL sets fail_type to
PRESUBMIT_FAILURE if the CQ attempt failed only because of failures in the
'chromium_presubmit' builder.

This is not a great solution, as we're still guessing as to whether a build
failed because of a missing LGTM.

Bug: 937519
Change-Id: Ia7e6f8964a3f0befc4e0f7a665bc7369f6686bec