qscheduler: add a WithFields method to scheduler.EventSink
Also, add an isCallback field to it.

Rationale: the value of isCallback should really be completely opaque to
any of the logic within qslib. Nevertheless, it still needs some way to
be passed in to the task events / metrics that qslib is emitting.

I considered a few alternatives:
 - Plumb the isCallback argument all the way through qslib into its
 metric emission.
 - Add isCallback as a context.Context value.

Both of them seemed worse than this approach.

BUG=940013

Change-Id: I2aeccf89db7806e5d2aaebc5ab9f15dbb9472fa0