Update patch set 2
Patch Set 2:

Thanks! This works for approval comments now, but this seems to also be broken for editing phases. I don't know for sure if the phase breakage is related as well, but we should fix that one as well.

Patch-set: 2
Reviewer: Gerrit User 1159599 <1159599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>