Update patch set 1
Patch Set 1: Code-Review+1 Commit-Queue+2

I investigated it some time ago. The problem is that git we are using is linking to a libc version not available on this Ubuntu. So our git fails to start. And our git_wrapper just does 🤷‍♂️ and falls back to system git, which is some prehistoric version that doesn't understand long authentication headers (remember we had to migrate to .gitcookies for that?). This manifests as auth errors from Gerrit.

Patch-set: 1
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1
Label: Commit-Queue=+2