Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> PTAL, demo here: http://22354-366b798-tainted-zhangtiff-rvg-dot-monorail-staging.appspot.com/ 
> 
> Basing this off of the latest mock in the doc for now. I figure it's okay if we end up needing to change this again later, as long as the indicator stays displayed in a way that people will notice it.

Sorry, what doc?
It seems to me that having the indicator the same color as the issue makes it easy to gloss over. I'd think it'd be better to have it with a yellow-ish background as it used to be.

Patch-set: 1
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>