Update patch set 3
Patch Set 3: Commit-Queue+2

(2 comments)

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> Ah, I see. Users glossing over it is not that big of a problem, since it's meant as a reassurance, right?
> 
> I still think it should be visually different from the rest of the header. Maybe we could do gray, or a light color?

I played with the color and some and wans't able to easily find a color that was different but also de-emphasized. So I'm sticking with green for now. We can change this in the future as we figure out more of where we want to go with restriction indicators, but for now, keeping it simple to unblock launch.

Patch-set: 3
Reviewer: Gerrit User 1159599 <1159599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2