[Findit v2] Process all rerun build results to narrow down regression ranges for each failure.
Uses results of all the rerun builds in an analysis to narrow down the regression ranges for each failure.
Then group the failures with the same regression range, so each group of failures can be run the next rerun build or share the same culprit.

This change only gets the failures and their updated regression range. Triggering next rerun build or finding culprit will be in https://chromium-review.googlesource.com/c/infra/infra/+/1594247.

Bug: 943716
Change-Id: I497a0ecb93181b8396bbfcec73378914f48db5e8