[tricium] Avoid building vendor deps, including those that use syscall
There are recent changes in transient dependencies of tricium (e.g.
dependencies of https://godoc.org/google.golang.org/grpc)
that mean that syscall is imported, e.g. in
https://github.com/googleapis/google-api-go-client/blob/master/transport/grpc/dial_socketopt.go#L22
and it's non-trivial to remove dependency on grpc related libraries.

After searching for a solution to this, one possible solution
is apparently to tell GAE to not build some packages.
https://ilikeorangutans.github.io/2016/05/22/google-appengine-go-and-vendoring/

I tried deploying a tainted version with this change, and it
apparently worked.

Change-Id: I9aeee0dd5b4892ffdf59113fd7414a1e3e71f5cb