[Monorail] Avoid offering autocomplete options that are already used.
For CCs and components, on the web component issue detail page, this
simply does not offer values that are already used.

For labels on the EZT and web components issue detail page, this
also applies the semantics of exclusive lable prefixes.  E.g., if
Milestone is an exclusive prefix and Milestone-1.0 has already
been used, then no other Milestone-* labels will be offered.

R=zhangtiff@chromium.org

Bug: monorail:5190
Change-Id: I8547d05d565770e6b717d58d033144db1ad63841