Update patch set 1
Patch Set 1:

There is no linear order between SkylabExecute and AutotestExecute.
It's pretty standard to sort such lists alphabetically. So if we don't, someone is likely to come cleanup ;)

If you want to keep this order, please add a comment to the code explaining that this is the order these steps are roughly used in.

Patch-set: 1
Reviewer: Gerrit User 1002639 <1002639@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>