Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > in https://bugs.chromium.org/p/chromium/issues/detail?id=978867#c8, you mentioned 10.14.4+, while this runs regardless of OS version. Is this safe to do prior to 10.14.4?
> > 
> > My thinking was that this should be a no-op at worst, 10.14.3 below and xcode10.1 and below...I don't have that configuration to confirm though, so is it possible for me to send this maybe a trybot that is running on 10.14.3 and running xcode10.1 to check?
> 
> Friendly ping?

We definitely don't have a way to exercise this on trybots prior to landing. If you don't have a 10.14.3 machine to confirm this on, it'd be good to try to keep the checkFirstLaunchStatus check on 10.14.3 & below if possible.

Patch-set: 1
Reviewer: Gerrit User 1181690 <1181690@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>