Monorail: remove cmp() builtin (reland)
This replaces usage of the cmp builtin, which is deprecated in Python 3,
with functools.total_ordering.

This is a reland of crrev.com/c/1690937. The reason the original CL failed
was because I forgot to update the attributes each class was checking. Lesson
learned - don't rely on unit tests to catch all errors.

Bug: monorail:5697
Change-Id: I28a26739ffd80043ceda240fbe1dc5a20bbc63db