Allow for additional tools install for xcode10.2+
Since xcode10.2, 10.2.1 and 11b5 may ask on runs after the first one to install
additional tools, we need to runFirstLaunch even if each run is
not the first run. At worst its a no-op of max 17ms (for me locally
at least) and at best it catches the subsequent/additional tool
install prompts for these version of xcode.

Bug:978867
Change-Id: Ib5443c2d87a83677d23f3cadd22b209667321ea9