Update patch set 12
Patch Set 12:

(9 comments)

> Patch Set 10:
> 
> (9 comments)
> 
> Awesome, thanks for working through my comments and splitting up this CL! The code is looking good to me though this CL does still seems to have the backend API changes in it.

PTAL at the latest patchset! Note: I know you suggested that I remove the selector for the object storage of starredIssues, but the set selector threw an undefined error when I did so, so I've left it in for now. Thanks for all of the help with this so far!

Patch-set: 12
Reviewer: Gerrit User 1349501 <1349501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>