[Monorail] Remove link to out-of-date Endpoints API design doc.
Let's remove this link because the relevant parts of the Endpoints API
design doc are already in api.md.  The other old info there probably
should not be shared with anyone who would need to request access, e.g.,
it has the usernames of former team members, and speculation on future
plans that turned out somewhat differently.

R=jojwang@chromium.org

Change-Id: I5345dfabb18d16a4d12bc08a8b3dcd8044f1cf8e