Update patch set 3
Patch Set 3:

I tested on monorail-dev (with a tiny settings.py change to enable it there) and it worked.  I will land this as after we are done with the current deployment, so as to not risk complicating that.

Patch-set: 3
Reviewer: Gerrit User 1163684 <1163684@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>