ipc_pubsub: add pubsub wrapper and implement subscribe
subscribe reads a channel of Messages, with supporting functions to set
up the channels it uses to communicate and to handle the direct
interface from receiving unwrapped Pubsub messages to placing wrapped
messages on the appropriate channel.

Testing package has a dummy channel which gives deterministic control
over what messages are received in what order, in order to test the
behavior of subscribe.

BUG=chromium:928535
TEST=internal/cmd/subscribe_test.go

Change-Id: Ice112bd6d35fb5e4e3a76f350b806d796eba29bc