[Monorail] Get user prefs when generating emails.
This just gets the UserPrefs objects for each address that we are
considering sending an email to.  Also-notify and project notification
addresses are just strings that may or may not correspond to accounts,
so an empty UserPrefs is used if there is no account.

In an upcoming CL, we will check these prefs and the issue restrictions
to decide if a link-only notification should be sent.

R=jojwang@chromium.org

Bug: monorail:2527
Change-Id: I5cabe56839baa7f426fe63b8f928b66a969002e8