[Monorail] Reduce concurrent default tasks to avoid DB replication lag.
I noticed that some recent comments were not showing up.
I checked DB replication lag and found that it started at 3:15pm,
which is the same time that a lot of recomputeDerivedFields tasks
were enqueued.  These tasks can do a lot of writes, but we don't
want to do so many writes per second that we cause DB replication lag.

R=jeffcarp@chromium.org

Change-Id: Ic517a865cd8f4894bf906495b3949d7fa5bfefe2