[monorail] Delete code unused to access dismissed cues.
These methods were used on the EZT pages before all cue card dismissal
calls were switched to SetUserPrefs().

Since April 2019, all dismissed cue information has been stored in the
UserPrefs table.  The advantage is that now we only load dismissed
cue information for the signed in user, not for all users that are
shown on the current page (e.g., all users who posted comments on
the current issue).  This simplifies the code and avoids a few SQL
queries.

R=jojwang@chromium.org

Bug: monorail:5934
Change-Id: Ie4ee57f974cd1eabb7be27675ef15c9cecf19db7