Show Federated Reference summaries in the UI
This CL modifies the FedRef classes to return more information about each
federated issue, in the format of an issue ref object. This CL expands that
information from just the open/closed status to the issue summary as well.

This CL should enable, once logged in, viewing the issue summary for
a Buganizer bug by hovering over it or opening the related issues modal.

Bug: monorail:6471
Change-Id: I06e569d70fb410a5e18f6dd385d6f29100620fac