Update patch set 2
Patch Set 2:

(2 comments)

I initially wanted to change these 500 to 404s because every case I saw was due to a patchset that was not present in the change. After reflecting on the feedback, it makes sense that we only return 404s in the case where gerrit itself is giving us 404s, for any other cases we should still return 500s (so that we get alerted if it's too frequent, and can then investigate).

I have also added a test for verifying these two behaviors. PTAL

Patch-set: 2