[cloudbuildhelper] Improve handing of `-canonical-tag :inputs-hash`.
When using regular tags, we check for a presence of a corresponding docker image
before anything else. This check was skipped when using `:inputs-hash` because
the exact value of `:inputs-hash` is not known before the docker context tarball
is built.

This was causing surprising and unnecessary rebuilds under some specific edge
case conditions.

This CL adds a check for existing docker images matching `:input-hash` canonical
tag after its exact value is known.

R=iannucci@chromium.org, tandrii@chromium.org

Change-Id: I0ca47235f2bbfa3809d99cf612ed330d6009e990