[gae_ts_mon] modernize gae_ts_mon enable djangomiddleware only in Python 2.
This CL includes the following changes.

(1) modernize gae_ts_mon
Note that this includes changes only for the python files that gae_ts_mon
directly owns. In other words, this CL doesn't include changes for modernizing
the python files under the symlinked folders/files, including
- common/,
- protobuf/,
- httplib2_utils.py,
- etc

(2) Enable gae_ts_mon.DjangoMiddleware only in Python 2.
It seems that gae_ts_mon.DjangoMiddleware is not used and needs to be rewritten.

cs.chromium.org shows that chromium_build and chromium_rietveld are the only
Infra apps using django, but chromium_build doesn't use
gae_ts_mon.DjangoMiddleware. chromium_rietveld uses the middleware, but,
it seems that the monitoring pipeline is broken. Looking at the following
dashboard, monitoring data points for chromium_rietveld have not been retained,
at least not in the past year.
: http://shortn/_8Y0bNkdsaK

Secondly, the DjangoMiddleware resolves the full path of the given view_func,
and reports HTTP monitoring data points with the path of the function object.
e.g., my_module.my_class.foo
HTTP traffic should be reported with the request path, such as /my_request,
not the python module/func/class path of the view func.

Also, there is no exact equivalent for im_func and im_class in python 3.
The code can be improved, like below

```
if sys.version_info.major == 3:
  if is_bound_method(view_func):
	  return '.'.join([...., view_func.__func__, ...])
  if ....
if sys.version_info.major == 2:
  return '.'.join([view_func.im_class.__module__, ...])
```

However, given that the middle ware code isn't really used anywhere effectively,
and the middleware should report HTTP traffic with the request path, instead,
it seems reasonable to limit the use of the current implementation of
gae_ts_mon.DjangoMiddleware to Py2. If there are needs to
fix the implementation, or support django-py3, then a separate CL should be
made for the work.

R=agable,iannucci

Bug: 1003185
Change-Id: I87771e3346e579176b7234f0d061a9db363c9a2f