[gae_ts_mon] Remove DjangoMiddleware
In https://crrev.com/c/1928352, it was found that
1. chrome_rietveld is the only app importing gae_ts_mon.DjangoMiddleware
in infra repo,
2. it doesn't look reasonable that the middleware reports HTTP traffic with
the full path of the python function object, e.g, my_module.my_class.my_func, and
3. the monitoring pipeline for chrome_rietveld is broken, and no HTTP monitoring
data has ever been reported for chrome_rietveld in the past year.

The usage of the middleware in chromium_rietveld has been removed by
https://crrev.com/c/1928645, and this CL is to remove the middleware from
gae_ts_mon, as the implementation doesn't get used by any app, and it needs
to be reimplemented to be used, anyways.

Bug: 1003185
Change-Id: I26205084d6815c75af7ed6a5aa3121ccad45971a
R=agable,iannucci