[recipe] Clean up usage of service accounts in recipes
The service account is a deprecated concept which exist in legacy build bot.
This is part of the effort to remove the depot_tools' cipd recipe module and
substitute it with the one in recipe engine. The cipd module in recipe engine
no longer has service account support. Hence, cleaning up all the current
usage in downstream repo.

R=iannucci

Bug: 875523
Change-Id: Ib7404d1e4eb3795ff51e24ef61911703c729927a