[gae_ts_mon] deprecate param cron_module for initialize() and remove handlers.app
This CL contains the following changes.

1. Deprecate gae_ts_mon.initialize takes param "cron_module".
The description says that it is to specify the name of the module handling
requests for /internal/cron/ts_mon/send.

However, it's incorrect information. In Appengine, cron.yaml specifies that
the name of a service to serve the HTTP traffic for a given cron path
: https://cloud.google.com/appengine/docs/standard/python/config/cronref#cron_job_definitions

Whichever value is passed on the param "cron_module", cron requests will be
routed to and served by any modules(services) matching with the target predicate
specified in cron.yaml. The param does make a difference, ts_mon report, though.
Only the module specified in the param will report the cron job requests into
the standard HTTP metrics, including /chrome/infra/http/server_response_status.

As a result, for example, if the value of "cron_module" doesn't match with
the configuration in cron.yaml, it causes tsmon to skip reporting
the success/failure of the cron job. However, the cron request will be handled,
anyways.

This CL deprecates and makes the param a noop.

2. Remove handlers.app.

handler.py creates a webapp2 instance in the module scope. As a result,
the webapp2 instance is automatically created, when an application imports
gae_ts_mon.

It has the following issues.
(a). Endpoint '/internal/cron/ts_mon/send' can be activated without or before
gae_ts_mon.initialize() invoked. It shouldn't.

(b) Unnecessarily dependency.
handlers.py is to manage the mapping of Appengine instances with task_num.
It is generic enough to be used by any HTTP framework, but the global var, app,
makes it hard to use handlers.py for other HTTP frameworks, such as Flask.

Intead of delcaring a webapp2 instance in the module scope, this CL add the
cron handler into the given webapp2 instance. As a result, the following
*backward-incompatible* change is made by this CL.

- Before this CL
Applications can import and invoke gae_ts_mon.initialize() without declaring
own webapp2 instance. For example,
```
  # there is no WSGI app for this app.
  gae_ts_mon.initialize(app=None)
```

- After this CL
Applications must declare and pass own webapp2 instance to use gae_ts_mon.
Even if the application has no business logic to handle HTTP traffic, the app
must declare a webapp2 instance with an empty router.

```
  app = webapp2.WSGIApplication()
  gae_ts_mon.initialize(app)
```

Even if gae_ts_mon allows None for param app, webapp2 is added as a dependency
to the application. There is no benefit of gae_ts_mon declaring an own webapp2
instance. It just causes code less intuitive, and makes it hard for gae_ts_mon
to support multiple HTTP frameworks.

Just in case there is any application using gae_ts_mon with
initialize(app=None), this CL raises an exception to warn that the app must
declare and pass an webapp2 instance to gae_ts_mon.initialize().


Change-Id: Ib746ecab7374e8a6da0089c03b3bfe59e0b6f821