Playing around with API modifying objects
Upon continued reflection, I think that the complex flexibility we want to provide with the combination of single (and batch) issue updating cannot be cleanly mapped to standard API methods.

This design emerged as I was continuing to struggle with elegent ways to handle
1. offering the API client control over when email is sent
2. adding a comment in association with one or more issue changes
3. uploading/removing attachment(s) in association with an issue change

---

I had the realization that most of our API clients will likely be experienced monorail users. Thus my hypothesis is that they will actually prefer an "update" API that more closely matches what is available in in the UI (for both single issue updates and batch issue updates).

----

Note also - I haven't thought through this in detail, but I wanted to go ahead and share something that represented the general direction.

Change-Id: If71ec0d69d523adfe93d3cd6350cc43513e38b3e