**DO NOT MERGE** Playing around with API modifying objects
Upon continued reflection, I think that the complex flexibility we want to provide with the combination of single (and batch) issue updating cannot be cleanly mapped to standard API methods.

This design emerged as I was continuing to struggle with elegent ways to handle
1. offering the API client control over when email is sent
2. adding a comment in association with one or more issue changes
3. uploading/removing attachment(s) in association with an issue change

---

I had the realization that most of our API clients will likely be experienced monorail users. Thus my hypothesis is that they will actually prefer an "update" API that more closely matches what is available in in the UI (for both single issue updates and batch issue updates).

Furthermore, for API clients who aren't familiar with the UI, we want to guide them to make changes to issue(s) in ways that will feel clean in the current monorail UX (i.e. batched changes with an associated comment in most cases; discretion on how many emails sent)
----

Note 1: I haven't thought through this in detail, but I wanted to go ahead and share something that represented the general direction.
Note 2: The pattern provided in 6.1 is likely still best for all cases other than Issue
Change-Id: If71ec0d69d523adfe93d3cd6350cc43513e38b3e