tricium-spellchecker: don't check footer flags in commit description
A little while ago, we enabled checking the whole
commit description, and since then have gotten a lot
of reports of false positives of people's names.

This CL would change it so that the last paragraph
(which contains footer flags) would be stripped before
checking so that no warnings are generated for footer
flag lines.

Bug: 1008712
Change-Id: I7cbc203bfb87eb1c77203257495f83aec9f05228