cros lab inventory: bugfixes of importing adapter
Bug 1: When the old specs (v1 format) doesn't have servo or rpm in
attributes, the code may return an pointer of uninitialized variable.

Bug2: For a labstation, the servo_host and servo_port referred in old
specs IS NOT the servo managed by this labstation. So don't set it to
`Servos` field.

BUG=chromium:962689
TEST=Ran the import tool

Change-Id: I2254cb36d4ef100d104f0fe0426a44d7d5bed3f6