[skylab] Factor out dims var into skylab command library.
As part of auditing the CTS allocation, we need to be able
to supply additional dimensions to skylab dut-list besides
model, board, and pool. dut-list.go lives in a different
submodule than lease-dut.go, so in order to share the dimsvar
functionality, it needs to be moved somewhere where all the
skylab commands can use it.

BUG=chromium:1069303
TEST=run tests locally AND ./skylab lease-dut -model garg -dims nonexistent_dim=74 AND confirm that task is created but fails with NO_RESOURCE.

Change-Id: I4b0390a3d4e749cbb64d64a8ee98b215fc18c9e1