SoM Disable Grouping: add cronjob to delete resolved alerts
Unlike AlertJSON that can be reused (i.e. when a builder failed, passed, and failed again, it is considered the same alert), for AlertJSONNonGrouping, we associate and alert with the first build it failed. Therefore we can expect the number of AlertJSONNongrouping will be much bigger than AlertJSON as they accumulate over time.

Moreover, once an AlertJSONNonGrouping is resolved, it cannot go back to "unresolved" state, because it will be consider a new alert.

Therefore, we should and we can flush those alerts.

I personally feel we can set the expiry duration to something short (like 1 day), but Joey thinks we may have some examine URL that links to a resolved alert, so we agree with the middleground of 1 week.
Bug:1443371
Change-Id: I2343b7ce4b6f1a26a844ddd4a2866a47a005c3f5