SoM: Reduce frequency of querying bug queue
We are getting rate limiting from Monorail. While we are increasing the rate limit from Monorail side, we should also reduce unnecessary calls to Monorail.

The bugqueue cron jobs are reponsible for a lot of calls to Monorail, and they are not really important anyway. They just refresh the bug data for some labels. They should not be called every minute.

For staging, we should run them even less frequently.

I make it 4 minutes so it does not overlap much with /_cron/annotations/refresh/, which also queries Monorail and run every 15 minutes.
Bug:1087689
Change-Id: I331f18e50594cd94ac5e5ddb2d9f2d9b55baa0b9