Update patch set 4
Patch Set 4:

Although I see the argument, I would still recommend we include parent.

For me the not mentioned big +  is consistency with AIP, so it will be easy for users.

I don't recommend we do any optimizing around it currently, but it leaves us open to it in the future for the cost of one upfront check.

Patch-set: 4
Tag: mailMessageId=<CAFdZGX5OT-rn-G2ykUEK=VHpC9gF97eOhN_CkMs_sZPOLPM4wg@mail.gmail.com>