[dirmeta] Fix wpt.notify inheritance
Proto3 does not distinguish between false and absence of the value.
This means that if dir `a` has wpt.notify=true and `a/b` has
wpt.notify=false explicitly, then after merging the two, wpt.notify is
still true.

Use BoolValue instead
https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/wrappers.proto#L101
This text-proto syntax is a awkward:
  wpt {
    notify { value: true }
  }
But it looks sane in JSON form:
  { "wpt": {"notify": true} }

The alternative is to use an enum with values (UNSPECIFIED, YES, NO).
Then the text proto syntax would be sane:
  wpt {
    notify: YES
  }
But JSON would look awkward:
  { "wpt": {"notify": "YES"} }

Bug: 1104246
Change-Id: I49c86ea9e7662a4e59b1b62252006376276ca114