[dirmeta] Add compute subcommand
Subcommand "compute" accepts a root, and 0 or more target directories.
For each target directory, it reads metadata for each directory on the
path from the root to the target.
It reads each directory at most once.
Then it computes inherited metadata for each target directory
and prints the mapping in the same format as "export" subcommand.

Also this fixes a bug that error was ignored in os.Walk function.


Change-Id: Id3097199fe5d11c6fe271cae9e9abf96565261da