[dirmeta] Implement Reduce
Reduce removes all redundant information from a Mapping.

This CL also removes Mapping.Expand() function.
It turned out that it is not useful in the Mapping struct,
because we need an entry for each directory existing on the file system.
For that we need to know the root path, which is not available
in the Mapping struct.

Bug: 1104246
Change-Id: I89e58a5caf9fb0f015deac58f724c0aa338ab740