Manual Repair: search by hostname component basic implementation
The component for search by hostname is loosely implemented here with
console logging. Next step is to implement authentication from the
frontend so that the backend RPC server will be able to verify
credentials and authenticate operations. A PRPC client is implemented
using @chopsui/prpc-client and is used to dispatch frontend actions.

BUG=chromium:1079495
TEST=npm i; npm run dev; make dev

Change-Id: Iff7eebdfed0af4d46277b99943f0cb46099ff836