exec-engine: Add skeleton of protobufs for the engine.
This defines a few core protobufs, and adds a Makefile with a rule to
regenerate the python modules for them.  The generated code is checked
in, consistent with elsewhere in the infra repository.  The proto
sources live in a proto/ subdirectory, and the generated files in a pb/
subdirectory.

This is not at all complete, but it's a reasonable starting point that
we'll build on in follow-up CLs.

Change-Id: I67a9f2486434a59b5ea359f3eaf516c237313455