Remove randomised test variants from pep425_test.
These are apparently to test that pep425TagSelector isn't dependent on
the order of the input tags. But this function is in fact dependent on
the order; it selects the maximum based on some criteria, but not every
field is checked.  So if two tags are identical in all checked fields,
the order determines which is selected.

Only two testcases have order-dependent behaviour. Both of them only
have two tags in the list. As far as I can tell, when this test was
written we got lucky with the default RNG seed and hit the identity
permutation for both randomised testcases (1/4 chance), which allowed
the tests to pass. You can make the test fail by setting a different
seed, or by adding a new testcase and perturbing the RNG state.

Change-Id: I1ff175251ff7029205d5af4016561a2f67412e6b