chromeperf: Define the WorkflowTemplate protobuf
This change introduces the WorkflowTemplate, which describes the kinds
of templates that will be defined in luci-config hosted files. This
template definition will be used in user interfaces to influence how we
build things like HTML forms and command-line utilities.

Protos that describe workflows will provide the set of supported inputs,
how those inputs are turned into options for creating workflow graphs,
and the Chrome Infra Auth groups that are allowed to see/use these
templates and the instantiated workflows.

The WorkflowTemplates service definition will only provide RPCs to get a
specific workflow template as well as list workflow templates with some
optional filtering.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1059667
Change-Id: I52a50a8be6dd0454e1c12be66dfb003566ac0e48