[infra_libs] Bump version from 2.1.0 to 2.2.0b1.
This release includes 4 changes that add experimental Py 2/3 support.
https://chromium-review.googlesource.com/c/infra/infra/+/2279396
https://chromium-review.googlesource.com/c/infra/infra/+/2279403
https://chromium-review.googlesource.com/c/infra/infra/+/2279405
https://chromium-review.googlesource.com/c/infra/infra/+/2532746

I want to release these changes to unblock Monorail / gae_ts_mon's
Python 3 migration, which depend on infra_libs through infra/ENV.

This is experimental (*b1) because the changes aren't fully tested
in Python 3, which would require updating infra/ENV to Python 3 or
replacing it with vpython. This is a larger effort planned for Q1.

Change-Id: Ibee62f784c69d839cb1d8ac146df4720725d0b68