[stable_version2] Remove non-inclusive language from comments.
We use the term "master" in three places to identify the main
git branch. Two of these cases are in comments. I am changing them
early before the name of the repo changes to "main". I am adding
a TODO comment to the final place where this name occurs and will
rename once the upstream project renames its main branch.

This is routine maintenance associated with the bug below.

BUG=b/179386397
TEST=does it compile.

Change-Id: I0fd6f63ea2444e73aacefd97674e74df5923a709