chromeperf: Pinpoint CLI (initial commit)
This change defines a Pinpoint command-line interface which interacts
with the Pinpoint gRPC API, hosted behind an API gateway that supports
OpenID Connect JWTs in the per-rpc request headers.

We'll be adding more subcommands as we fill out the functionality on the
service-side support for those kinds of commands as well.

In this change we also add dependencies to allow us to use go-oidc to
simplify obtaining an ID token.

--------------------------------------------------------------------------
Updated repos:
https://chromium.googlesource.com/external/github.com/coreos/go-oidc.git/+log/fbec1bdfed4d..a93f71fdfe73
https://chromium.googlesource.com/external/github.com/square/go-jose.git/+log/3a5ee095dcb5..748cb1f5affb
--------------------------------------------------------------------------

R=abennetts@google.com, fancl@chromium.org, nodir@chromium.org

Bug: chromium:1059667
Change-Id: I11e77ee2d353d78df6b43d9e5b6159e10a86ce3b