[rts] Remove always-run tests
Earlier we considered filtering tests at the compilation stage. This
requires always running tests that might contain main() function or
might be a dependency of another file.
Since then, we decided to do filtering in test runtime, so we no longer
need this extra filtering, and specifically the test file denylist.
Remove this functionality.

Bug: 1090668
Change-Id: I347bdf314f1634691f567156e0faf14b6fb3dbe9