Simplify PinpointClient setup; support insecure mode for localhost.
I didn't quite see the need to have commands separately init the factory
and then get a pinpoint client, so I made it so they didn't need to.
This CL doesn't change any caching aspects of the factory, and callers
can safely call pinpointClient in parallel if they wanted to.

I went with auto-detecting whether we should use insecure mode instead
of adding a flag, though admittedly it's a little hacky since I just
hardcoded "localhost" into the check rather than trying to do something
smarter which would cover more scenarios. Let me know if you'd like this
to be a smarter, or if you think a flag to enable insecure mode is best.

Change-Id: I73c262326a3992034b5ed3ba05965d064a3171f9