cli: Very basic support for list-jobs.
I wrote this as a way to get more familiar with the codebase, before I
realized that ListJobs was not implemented on the server... I figured
I'd checkpoint the work anyway instead of sitting on it, but feel free
to request that I hold back on this for now.


Simplify PinpointClient setup; support insecure mode for localhost.

I didn't quite see the need to have commands separately init the factory
and then get a pinpoint client, so I made it so they didn't need to.
This CL doesn't change any caching aspects of the factory, and callers
can safely call pinpointClient in parallel if they wanted to.

I went with auto-detecting whether we should use insecure mode instead
of adding a flag, though admittedly it's a little hacky since I just
hardcoded "localhost" into the check rather than trying to do something
smarter which would cover more scenarios. Let me know if you'd like this
to be a smarter, or if you think a flag to enable insecure mode is best.

Change-Id: Ib0ae121e7ec2dc6ff098c2cc914bff88d1433cac