cli: support get-job and wait-job
Specific output is subject to discussion; I picked arbitrarily based on
what I might want to see if it were me, but I've have not yet used
pinpoint :)

We might want to support more convenient forms of the --name, e.g.
perhaps we should be prepending to "jobs/legacy-" instead of requiring
users to do it for us. Please let me know what your thoughts are here.

We might want to support a `--wait` option for the commands which submit
jobs to execute; in that case, we should refactor the logic of
waitJobs.Run to be more reusable, though perhaps it does not need to
happen in this CL.

In my testing, I realized that wait-job might loop long enough that the
JWT expires; I dropped a TODO to remind myself to address that, though I
hope to do so next week.


cli: Very basic support for list-jobs.

I wrote this as a way to get more familiar with the codebase, before I
realized that ListJobs was not implemented on the server... I figured
I'd checkpoint the work anyway instead of sitting on it, but feel free
to request that I hold back on this for now.

Change-Id: I3275e13d0f75f9f0e8d80d872ce1da2410d824c4